pyproject.toml
[tool.poetry]
name = "doc-label"
version = "0.1.0"
description = ""
authors = ["Narayan Bandodker <narayanband1356@gmail.com>"]
readme = "README.md"


[tool.poetry.scripts]
# CLI script
doc-label = "doc_label.cli:main"


[tool.poetry.dependencies]
# Main dependencies

python = "^3.11"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
platformdirs = "^4.2.1"
fastapi = "^0.111.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tqdm = "^4.66.4"
sqlalchemy = "^2.0.30"
aiosqlite = "^0.20.0"
aiodecorators = "^0.2.1"
pyyaml = "^6.0.2"
pymupdf = "^1.24.10"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
unidecode = "^1.3.8"
shapely = "^2.0.6"
numpy = "^2.1.1"
pandas = "^2.2.2"
lxml = "^5.3.0"


[tool.poetry.group.test]
# Testing dependencies
optional = true

[tool.poetry.group.test.dependencies]
# Pytest for VSCode test-suite compatibility
pytest = "^8.1.1"


[tool.poetry.group.dev]
# Development dependencies
optional = true

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.0"
pre-commit = "^3.7.0"
watchfiles = "^0.21.0"
uvicorn = "^0.29.0"


[tool.black]
# Tool configurations

## Black (formatter)
line-length = 120
target-version = ['py311']
include = '\.pyi?$'


[tool.flake8]
## Flake8 (PEP-8 linter)
ignore = "E203, E266, E501, W503, F403, F401"
exclude = ".git, __pycache__, venv, .vscode, .mypy_cache, .pytest_cache, build, dist"
max-line-length = 120
max-complexity = 18
select = "B,C,E,F,W,T4,B9"
accept-encodings = "utf-8"
inline-quotes = "single"
max-linenumber = 500
multiline-quotes = "double"


[tool.pyright]
## Pyright (Static type checker)
venvPath = "."
venv = "venv"
reportShadowedImports=false


[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
